backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-ckeditor==6.7.3
django-cors-headers==4.9.0
django-email-users==2.1.0
django-js-asset==3.1.2
djangorestframework==3.16.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
pillow==11.3.0
psycopg2==2.9.10
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
django-jazzmin
django-filter==24.3
reportlab==4.2.5
